requirements.txt
python-frontmatter==1.0.0
markdown2==2.4.10
jinja2==3.1.2 